Move `SyncModuleConfig` class to `SyncConfigModule`
