+++ app/src/main/java/org/simple/clinic/sync/SyncConfigModule.kt
@@ -3,6 +3,7 @@ package org.simple.clinic.sync
 import dagger.Module
 import dagger.Provides
 import org.simple.clinic.remoteconfig.ConfigReader
+import java.util.Locale
 import javax.inject.Named
 
 @Module
@@ -31,5 +32,41 @@ class SyncConfigModule {
   @Provides
   fun syncModuleConfig(reader: ConfigReader): SyncModuleConfig {
     return SyncModuleConfig.read(reader)
+  }
+
+  data class SyncModuleConfig(
+      val frequentSyncBatchSize: BatchSize,
+      val dailySyncBatchSize: BatchSize
+  ) {
+
+    companion object {
+
+      fun read(reader: ConfigReader): SyncModuleConfig {
+        val frequentConfigString = reader.string("syncmodule_frequentsync_batchsize", default = "large")
+
+        val frequentBatchSize = when (frequentConfigString.toLowerCase(Locale.ROOT)) {
+          "verysmall" -> BatchSize.VERY_SMALL
+          "small" -> BatchSize.SMALL
+          "medium" -> BatchSize.MEDIUM
+          "large" -> BatchSize.LARGE
+          else -> BatchSize.MEDIUM
+        }
+
+        val dailyConfigString = reader.string("syncmodule_dailysync_batchsize", default = "large")
+
+        val dailyBatchSize = when (dailyConfigString.toLowerCase(Locale.ROOT)) {
+          "verysmall" -> BatchSize.VERY_SMALL
+          "small" -> BatchSize.SMALL
+          "medium" -> BatchSize.MEDIUM
+          "large" -> BatchSize.LARGE
+          else -> BatchSize.MEDIUM
+        }
+
+        return SyncModuleConfig(
+            frequentSyncBatchSize = frequentBatchSize,
+            dailySyncBatchSize = dailyBatchSize
+        )
+      }
+    }
   }
 }
